Move test build actions to separate package
