+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/actions/FetchCustomModelForEachProjectInParallel.java
@@ -14,38 +14,38 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.cc.impl.isolated;
+package org.gradle.internal.cc.impl.actions;
 
 import org.gradle.internal.cc.impl.fixtures.SomeToolingModel;
 import org.gradle.tooling.BuildAction;
 import org.gradle.tooling.BuildController;
+import org.gradle.tooling.model.gradle.BasicGradleProject;
+import org.gradle.tooling.model.gradle.GradleBuild;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class FetchCustomModelForSameProjectInParallel implements BuildAction<List<SomeToolingModel>> {
-
+public class FetchCustomModelForEachProjectInParallel implements BuildAction<List<SomeToolingModel>> {
     @Override
     public List<SomeToolingModel> execute(BuildController controller) {
         List<FetchModelForProject> actions = new ArrayList<>();
-        for (int i = 1; i <= 2; i++) {
-            actions.add(new FetchModelForProject("nested-" + i));
+        GradleBuild buildModel = controller.getBuildModel();
+        for (BasicGradleProject project : buildModel.getProjects()) {
+            actions.add(new FetchModelForProject(project));
         }
         return controller.run(actions);
     }
 
     private static class FetchModelForProject implements BuildAction<SomeToolingModel> {
+        private final BasicGradleProject project;
 
-        private final String id;
-
-        private FetchModelForProject(String id) {
-            this.id = id;
+        public FetchModelForProject(BasicGradleProject project) {
+            this.project = project;
         }
 
         @Override
         public SomeToolingModel execute(BuildController controller) {
-            System.out.println("Executing " + id);
-            return controller.findModel(SomeToolingModel.class);
+            return controller.findModel(project, SomeToolingModel.class);
         }
     }
 }
