+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/actions/FetchCustomModelForTargetProject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.cc.impl.isolated;
+package org.gradle.internal.cc.impl.actions;
 
 import org.gradle.internal.cc.impl.fixtures.SomeToolingModel;
 import org.gradle.tooling.BuildAction;
@@ -22,30 +22,29 @@ import org.gradle.tooling.BuildController;
 import org.gradle.tooling.model.gradle.BasicGradleProject;
 import org.gradle.tooling.model.gradle.GradleBuild;
 
-import java.util.ArrayList;
-import java.util.List;
+public class FetchCustomModelForTargetProject implements BuildAction<SomeToolingModel> {
 
-public class FetchCustomModelForEachProjectInParallel implements BuildAction<List<SomeToolingModel>> {
-    @Override
-    public List<SomeToolingModel> execute(BuildController controller) {
-        List<FetchModelForProject> actions = new ArrayList<>();
-        GradleBuild buildModel = controller.getBuildModel();
-        for (BasicGradleProject project : buildModel.getProjects()) {
-            actions.add(new FetchModelForProject(project));
-        }
-        return controller.run(actions);
+    private final String targetProject;
+
+    public FetchCustomModelForTargetProject(String targetProject) {
+        this.targetProject = targetProject;
     }
 
-    private static class FetchModelForProject implements BuildAction<SomeToolingModel> {
-        private final BasicGradleProject project;
-
-        public FetchModelForProject(BasicGradleProject project) {
-            this.project = project;
+    @Override
+    public SomeToolingModel execute(BuildController controller) {
+        GradleBuild buildModel = controller.getBuildModel();
+        for (BasicGradleProject project : buildModel.getProjects()) {
+            if (targetProject.equals(project.getBuildTreePath())) {
+                return controller.getModel(project, SomeToolingModel.class);
+            }
         }
-
-        @Override
-        public SomeToolingModel execute(BuildController controller) {
-            return controller.findModel(project, SomeToolingModel.class);
+        for (GradleBuild editableBuild : buildModel.getEditableBuilds()) {
+            for (BasicGradleProject project : editableBuild.getProjects()) {
+                if (targetProject.equals(project.getBuildTreePath())) {
+                    return controller.getModel(project, SomeToolingModel.class);
+                }
+            }
         }
+        return null;
     }
 }
