+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/actions/FetchPartialCustomModelForEachProject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.cc.impl.isolated;
+package org.gradle.internal.cc.impl.actions;
 
 import org.gradle.internal.cc.impl.fixtures.SomeToolingModel;
 import org.gradle.tooling.BuildAction;
@@ -25,24 +25,17 @@ import org.gradle.tooling.model.gradle.GradleBuild;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FetchCustomModelForEachProjectInTree implements BuildAction<List<SomeToolingModel>> {
+public class FetchPartialCustomModelForEachProject implements BuildAction<List<String>> {
     @Override
-    public List<SomeToolingModel> execute(BuildController controller) {
-        List<SomeToolingModel> result = new ArrayList<>();
+    public List<String> execute(BuildController controller) {
+        List<String> result = new ArrayList<>();
         GradleBuild buildModel = controller.getBuildModel();
-        collectModelsForProjects(controller, result, buildModel);
-        for (GradleBuild build : buildModel.getEditableBuilds()) {
-            collectModelsForProjects(controller, result, build);
-        }
-        return result;
-    }
-
-    private void collectModelsForProjects(BuildController controller, List<SomeToolingModel> result, GradleBuild buildModel) {
         for (BasicGradleProject project : buildModel.getProjects()) {
             SomeToolingModel model = controller.findModel(project, SomeToolingModel.class);
             if (model != null) {
-                result.add(model);
+                result.add(model.getMessage());
             }
         }
+        return result;
     }
 }
