+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/isolated/IsolatedProjectsToolingApiPhasedBuildActionIntegrationTest.groovy
@@ -16,6 +16,9 @@
 
 package org.gradle.internal.cc.impl.isolated
 
+import org.gradle.internal.cc.impl.actions.FetchCustomModelForEachProject
+import org.gradle.internal.cc.impl.actions.FetchCustomModelForTargetProject
+import org.gradle.internal.cc.impl.actions.FetchPartialCustomModelForEachProject
 import org.gradle.internal.cc.impl.fixtures.SomeToolingModel
 
 class IsolatedProjectsToolingApiPhasedBuildActionIntegrationTest extends AbstractIsolatedProjectsToolingApiIntegrationTest {
